Refactor Table components
Ticket: #372
